Add password confirmation and try out anynines
FIX: Now passwords have to be confirmed. The steps were rather simple.
Just adding the field in the view and whitelisting it in the 'signup_
params' method did the trick. Easy to undo.

Then I tried out anynines, didn't work. However the config file
(manifest.yml.v6) is still there. But for now it's not going to be used.