STYLE: Optimzing welcome page and css
Problem:
The backgound image only fitted because I added a padding the inversive
ratio (height/width) of the image and setting height to 0
Biggest problem with this was: The heading could not be verticaly
centered in the jumbotron

Change:
Now the approach is to set the height by em (in that case 44em),
however still not perfect.

Changes that came along with that:

- flash messages on the welcome page should not move the image
so I changed the application.css to only include the same directory,
but not subdirectories.
The new subdirectory holds css that should be nonincluded by default.
(They must be included via assets.rb)

Minor changes:
- included a test site to poiint'/try' to pages#try
- welcome heading is now sort of centered (value is top: 25%)