Add searches to app
Some basic code was already done(e.g. the database).
For the rest the offer controller and views where copied and edited.
So basically it works the same as the offers work.
The future work related to offers/searches is:
Make them correlate on creation and focus the app more on searches.