Make app no-script compatible
Especially deletion is the problem when not using js.

Deletion without js was already implemented for offers&searches even
tough it wasn't that important, however a big nuisance was when
logging out: (and still having js disabled)An error occured and you
could not log out(due to having no route fot GET /logout).

The only trade-off: Now there is a button in the navbar which has to
be consistent with the style of the app