Fix redirect after login
If an user tries to visit a page he needs to be authenticated for to
visit it, he gets redirected to the login page and the requested url
gets saved in the session cookie. However that value stayed in that
session as long as the user was logged in.
This commit changes this behaviour so that it gets deleted as soon as the user logs in.