Change redirect feature to use flash
The redirect feature which redirects users, who wanted to visit
a restricted site after a successful login to that same page.
This commit switches the method from session based to flash based,
in order to use the flash functionality to destroy itself (whereas
in the session based approach it had to be manually deleted).
However it seems it didn't change much in terms of simplicity
and clarity, so one might as well revert it.