MAJOR Add remember_me feature
Add all the necessary methods and tests and integrate it into the login
view.
One thing to note:
If users check the box on one computer and login on an other one without
checking the box, the first computer is still 'remembered',
i.e. the remember_digest of that user is untouched